recipes/requirements.txt
Django==2.1.7
djangorestframework==3.9.2
psycopg2==2.7.7
pillow==5.4.1
